Refactor lambda to handle a single file (WIP)
